+++ core/rust.module_settings/src/common_columns.rs
@@ -9,6 +9,7 @@ pub fn created_at() -> Column {
     Column {
         id: "created_at",
         name: "Created At",
+        description: "The time the record was created.",
         column_type: ColumnType::new_scalar(InnerColumnType::TimestampTz {}),
         nullable: false,
         unique: false,
@@ -25,6 +26,7 @@ pub fn created_by() -> Column {
     Column {
         id: "created_by",
         name: "Created By",
+        description: "The user who created the record.",
         column_type: ColumnType::new_scalar(InnerColumnType::String {
             min_length: None,
             max_length: None,
@@ -46,6 +48,7 @@ pub fn last_updated_at() -> Column {
     Column {
         id: "last_updated_at",
         name: "Last Updated At",
+        description: "The time the record was last updated.",
         column_type: ColumnType::new_scalar(InnerColumnType::TimestampTz {}),
         ignored_for: vec![OperationType::Create, OperationType::Update],
         secret: None,
@@ -62,6 +65,7 @@ pub fn last_updated_by() -> Column {
     Column {
         id: "last_updated_by",
         name: "Last Updated By",
+        description: "The user who last updated the record.",
         column_type: ColumnType::new_scalar(InnerColumnType::String {
             min_length: None,
             max_length: None,